host: uhd::transport::get_if_addrs() should only populate the broadcast address if available from getifaddrs().
It looks like Linux (and perhaps also others) populate the ifa_broadaddr field of the ifaddrs structure even if the
IFF_BROADCAST flags is not set in ifa_flags field. This is the case with the lo (loopback) interface where 127.0.0.1 is
returned even if the broadcast flag is not set.
Solaris does not show this behaviour .